Add LLM question command
The command

```
puddle ask "this is a prompt"
```

Will return the LLM content.

Currently the model is set at gpt-4, the parser is set as the string
identity parser, and the role is set to be a museum curator.

I am reconsidering whether I actually want to use langchain. I think
they have a bit too much layering.

So next up, I'll remove the langchain deps and then start implementing
the vector database integration.